+++ Include/Goblin/User/Interface/Drawable.hpp
@@ -102,75 +102,3 @@ public:
 		getTooltip = ntooltip;
 	}
 };
-
-/**
- * Abstract class
- */
-class Scrollable
-{
-
-public:
-
-	// Destructor
-
-	virtual ~Scrollable()
-	{
-	}
-
-	virtual void Draw(int x, int y, int scroll, int width, int height, TCODConsole*) = 0;
-
-	virtual int TotalHeight() = 0;
-
-	virtual MenuResult Update(int x, int y, bool clicked, TCOD_key_t key)
-	{
-		return NOMENUHIT;
-	}
-
-	virtual void GetTooltip(int x, int y, Tooltip* tooltip)
-	{
-	};
-};
-
-class UIContainer : public Drawable
-{
-
-protected:
-
-	std::vector<Drawable*> components;
-
-public:
-
-	UIContainer(std::vector<Drawable*> ncomponents, int nx, int ny, int nwidth, int nheight) :
-			Drawable(nx, ny, nwidth, nheight), components(ncomponents)
-	{
-	}
-
-	~UIContainer();
-
-	void AddComponent(Drawable* component);
-
-	virtual void Draw(int, int, TCODConsole*);
-
-	virtual MenuResult Update(int, int, bool, TCOD_key_t);
-
-	virtual void GetTooltip(int, int, Tooltip*);
-};
-
-class Panel : public Drawable
-{
-
-public:
-
-	Panel(int nwidth, int nheight) :
-			Drawable(0, 0, nwidth, nheight)
-	{
-	}
-
-	void ShowModal();
-
-	virtual void Open();
-
-	virtual void Close();
-
-	virtual void selected(int);
-};
